requirements.txt
crowdstrike-falconpy>=1.6.0
python-dotenv>=1.0.0
colorama>=0.4.6
